chore(webpack): split client and demo configuration
